+++ gui/pulsed/pulsed_maingui.py
@@ -27,6 +27,7 @@ import datetime
 
 from core.module import Connector, StatusVar
 from core.util import units
+from core.util.helpers import natural_sort
 from gui.colordefs import QudiPalettePale as palette
 from gui.fitsettings import FitSettingsDialog
 from gui.guibase import GUIBase
@@ -950,7 +951,7 @@ class PulsedMeasurementGui(GUIBase):
         for cfg in self.pulsedmasterlogic().pulse_generator_constraints.activation_config.values():
             for ach in (chnl for chnl in cfg if chnl.startswith('a')):
                 analog_channels.add(ach)
-        analog_channels = sorted(analog_channels, key=lambda chnl: int(chnl.split('ch')[1]))
+        analog_channels = natural_sort(analog_channels)
 
         for i, chnl in enumerate(analog_channels, 1):
             self._analog_chnl_setting_widgets[chnl] = (
@@ -990,7 +991,7 @@ class PulsedMeasurementGui(GUIBase):
         for cfg in self.pulsedmasterlogic().pulse_generator_constraints.activation_config.values():
             for dch in (chnl for chnl in cfg if chnl.startswith('d')):
                 digital_channels.add(dch)
-        digital_channels = sorted(digital_channels, key=lambda chnl: int(chnl.split('ch')[1]))
+        digital_channels = natural_sort(digital_channels)
 
         for i, chnl in enumerate(digital_channels, 1):
             self._digital_chnl_setting_widgets[chnl] = (
@@ -1062,7 +1063,7 @@ class PulsedMeasurementGui(GUIBase):
         editor.
         """
         # create all GUI elements and check all boxes listed in the methods to show
-        for method_name in sorted(self.pulsedmasterlogic().generate_methods):
+        for method_name in natural_sort(self.pulsedmasterlogic().generate_methods):
             # create checkboxes for the config dialogue
             name_checkbox = 'checkbox_' + method_name
             setattr(self._pm_cfg, name_checkbox, QtWidgets.QCheckBox(self._pm_cfg.scrollArea))
@@ -1154,10 +1155,8 @@ class PulsedMeasurementGui(GUIBase):
                 widget.setObjectName('global_param_' + param)
                 widget.setSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
                 widget.addItem('')
-                widget.addItems(sorted(self.pulsedmasterlogic().digital_channels,
-                                       key=lambda chnl: int(chnl.split('ch')[-1])))
-                widget.addItems(sorted(self.pulsedmasterlogic().analog_channels,
-                                       key=lambda chnl: int(chnl.split('ch')[-1])))
+                widget.addItems(natural_sort(self.pulsedmasterlogic().digital_channels))
+                widget.addItems(natural_sort(self.pulsedmasterlogic().analog_channels))
                 index = widget.findText(value)
                 if index >= 0:
                     widget.setCurrentIndex(index)
@@ -1235,7 +1234,7 @@ class PulsedMeasurementGui(GUIBase):
         self._pm.method_param_widgets = dict()
 
         method_params = self.pulsedmasterlogic().generate_method_params
-        for method_name in sorted(self.pulsedmasterlogic().generate_methods):
+        for method_name in natural_sort(self.pulsedmasterlogic().generate_methods):
             # Create the widgets for the predefined methods dialogue
             # Create GroupBox for the method to reside in
             groupBox = QtWidgets.QGroupBox(self._pm)
@@ -1370,12 +1369,10 @@ class PulsedMeasurementGui(GUIBase):
             self._pgs.gen_sample_freq_DSpinBox.setValue(settings_dict['sample_rate'])
         if 'activation_config' in settings_dict:
             config_name = settings_dict['activation_config'][0]
-            digital_channels = sorted(
-                (ch for ch in settings_dict['activation_config'][1] if ch.startswith('d')),
-                key=lambda channel: int(channel.split('ch')[-1]))
-            analog_channels = sorted(
-                (ch for ch in settings_dict['activation_config'][1] if ch.startswith('a')),
-                key=lambda channel: int(channel.split('ch')[-1]))
+            digital_channels = natural_sort(
+                (ch for ch in settings_dict['activation_config'][1] if ch.startswith('d')))
+            analog_channels = natural_sort(
+                (ch for ch in settings_dict['activation_config'][1] if ch.startswith('a')))
             index = self._pgs.gen_activation_config_ComboBox.findText(config_name)
             self._pgs.gen_activation_config_ComboBox.setCurrentIndex(index)
             digital_str = str(digital_channels).strip('[]').replace('\'', '').replace(',', ' |')
@@ -1774,7 +1771,7 @@ class PulsedMeasurementGui(GUIBase):
         @param block_dict:
         @return:
         """
-        block_names = sorted(block_dict)
+        block_names = natural_sort(block_dict)
         # Check if a block has been added. In that case set the current index to the new one.
         # In all other cases try to maintain the current item and if it was removed, set the first.
         text_to_set = None
@@ -1803,7 +1800,7 @@ class PulsedMeasurementGui(GUIBase):
         @param ensemble_dict:
         @return:
         """
-        ensemble_names = sorted(ensemble_dict)
+        ensemble_names = natural_sort(ensemble_dict)
         # Check if an ensemble has been added. In that case set the current index to the new one.
         # In all other cases try to maintain the current item and if it was removed, set the first.
         text_to_set = None
@@ -2088,7 +2085,7 @@ class PulsedMeasurementGui(GUIBase):
         @param sequence_dict:
         @return:
         """
-        sequence_names = sorted(sequence_dict)
+        sequence_names = natural_sort(sequence_dict)
         # Check if a sequence has been added. In that case set the current index to the new one.
         # In all other cases try to maintain the current item and if it was removed, set the first.
         text_to_set = None