+++ hardware/awg/tektronix_awg70k.py
@@ -32,6 +32,7 @@ from lxml import etree as ET
 
 from core.module import Base, ConfigOption
 from core.util.modules import get_home_dir
+from core.util.helpers import natural_sort
 from interface.pulser_interface import PulserInterface, PulserConstraints
 
 
@@ -340,7 +341,7 @@ class AWG70K(Base, PulserInterface):
         # determine active channels
         activation_dict = self.get_active_channels()
         active_channels = {chnl for chnl in activation_dict if activation_dict[chnl]}
-        active_analog = sorted(chnl for chnl in active_channels if chnl.startswith('a'))
+        active_analog = natural_sort(chnl for chnl in active_channels if chnl.startswith('a'))
 
         # Sanity check of channel numbers
         if active_channels != set(analog_samples.keys()).union(set(digital_samples.keys())):
@@ -439,7 +440,7 @@ class AWG70K(Base, PulserInterface):
                                'present in device memory.'.format(name, waveform_tuple))
                 return -1
 
-        active_analog = sorted(chnl for chnl in self.get_active_channels() if chnl.startswith('a'))
+        active_analog = natural_sort(chnl for chnl in self.get_active_channels() if chnl.startswith('a'))
         num_tracks = len(active_analog)
         num_steps = len(sequence_parameter_list)
 
@@ -479,8 +480,7 @@ class AWG70K(Base, PulserInterface):
                                    '"{1}".'.format(seq_step.go_to, num_steps))
                     return -1
             # Set flag states
-            if len(seq_step.flag_trigger) + len(seq_step.flag_high) > 0:
-                self.sequence_set_flags(name, step, seq_step.flag_trigger, seq_step.flag_high)
+            self.sequence_set_flags(name, step, seq_step.flag_trigger, seq_step.flag_high)
 
         # Wait for everything to complete
         while int(self.query('*OPC?')) != 1:
@@ -497,7 +497,7 @@ class AWG70K(Base, PulserInterface):
         except visa.VisaIOError:
             query_return = None
             self.log.error('Unable to read waveform list from device. VisaIOError occured.')
-        waveform_list = sorted(query_return.split(',')) if query_return else list()
+        waveform_list = natural_sort(query_return.split(',')) if query_return else list()
         return waveform_list
 
     def get_sequence_names(self):
@@ -585,7 +585,7 @@ class AWG70K(Base, PulserInterface):
 
         # Get all active channels
         chnl_activation = self.get_active_channels()
-        analog_channels = sorted(
+        analog_channels = natural_sort(
             chnl for chnl in chnl_activation if chnl.startswith('a') and chnl_activation[chnl])
 
         # Check if all channels to load to are active
@@ -627,7 +627,7 @@ class AWG70K(Base, PulserInterface):
 
         # Get all active channels
         chnl_activation = self.get_active_channels()
-        analog_channels = sorted(
+        analog_channels = natural_sort(
             chnl for chnl in chnl_activation if chnl.startswith('a') and chnl_activation[chnl])
 
         # Check if number of sequence tracks matches the number of analog channels
@@ -1462,7 +1462,7 @@ class AWG70K(Base, PulserInterface):
             for config in configs.values():
                 if len(largest_config) < len(config):
                     largest_config = config
-        return sorted(largest_config)
+        return natural_sort(largest_config)
 
     def _get_all_analog_channels(self):
         """